# packages/dashboard/openapi.yaml
openapi: 3.1.0
info:
  title: Aurora Dashboard API
  version: 0.3.0
  description: >
    Admin-plane API for aurora.local. Same-origin, cookie-authenticated
    (session cookie set by POST /auth/login). All paths below are served
    under the `/api` prefix by the Spring Boot backend; the Vite dev
    server proxies `/api` to :8090.


    This spec is hand-authored from the typed frontend `src/api/*`
    modules and is the source of truth the MSW mock layer answers to.
    Field casing is deliberately mixed: most bodies are camelCase, but a
    handful of endpoints emit snake_case verbatim from the backend
    (onboarding draft/plan/launch, services, containers logs, security
    dismissals, audit). Casing here matches the wire, not a convention.
servers:
  - url: /api
    description: Same-origin, behind Caddy / Vite proxy.
...
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Log in; sets the session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Logged in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401':
          description: Bad credentials.
  /auth/me:
    get:
      tags: [auth]
      summary: The signed-in user, or 401
      description: >
        Sibling of /auth/session. Public in SecurityConfig so the SPA can
        ask before it knows whether it has a session.
      responses:
        '200':
          description: The current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { description: Not signed in. }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: AURORA_SESSION
...
  schemas:
    Session:
      type: object
      required: [authenticated, username, passkeyEnrolled, tz]
      properties:
        authenticated: { type: boolean }
        username: { type: [string, 'null'] }
        passkeyEnrolled: { type: boolean }
        tz: { type: [string, 'null'] }
...